Refactor MFA e2e suite into reusable helpers
Extract the protocol and client machinery out of mfa.spec.ts into focused
modules under test/e2e/utils (totp, keycloak-admin, mail-protocol, oauth,
mfa) so the spec reads as scenarios rather than test infrastructure.
